+++ R/AutoRanger.R
@@ -31,7 +31,9 @@ AutoRanger = R6Class(
         properties = "large_data_sets",
         task_types = c("classif", "regr"),
         packages = c("mlr3", "mlr3learners", "ranger"),
-        devices = "cpu"
+        devices = "cpu",
+        n_cpu = 1L,
+        n_gpu = 0L
       )
     },