+++ R/AutoTabPFN.R
@@ -33,7 +33,9 @@ AutoTabPFN = R6Class(
         properties = character(0),
         task_types = c("classif", "regr"),
         packages = c("mlr3", "mlr3extralearners", "callr"),
-        devices = c("cpu", "cuda")
+        devices = c("cpu", "cuda"),
+        n_cpu = 1L,
+        n_gpu = 1L
       )
     },