+++ R/AutoXgboost.R
@@ -30,7 +30,9 @@ AutoXgboost = R6Class(
         properties = c("internal_tuning", "large_data_sets"),
         task_types = c("classif", "regr"),
         packages = c("mlr3", "mlr3learners", "xgboost"),
-        devices = c("cpu", "cuda")
+        devices = c("cpu", "cuda"),
+        n_cpu = 1L,
+        n_gpu = 0L
       )
     },