+++ R/LearnerAuto.R
@@ -84,6 +84,8 @@ LearnerAuto = R6Class(
         learner_timeout         = p_int(lower = 1L, init = 900L, tags = c("train", "super")),
         measure                 = p_uty(tags = c("train", "super"), custom_check = crate({function(x) check_r6(x, "Measure")})),
         memory_limit            = p_int(lower = 1L, init = 32000L, tags = c("train", "catboost", "extra_trees", "fastai", "ft_transformer", "lightgbm", "mlp", "ranger", "resnet", "tabpfn", "xgboost")),
+        n_cpu                   = p_uty(tags = c("train", "super"), custom_check = crate({function(x) check_integerish(x, lower = 1L, any.missing = FALSE, names = "unique", null.ok = TRUE)})),
+        n_gpu                   = p_uty(tags = c("train", "super"), custom_check = crate({function(x) check_integerish(x, lower = 0L, upper = 1L, any.missing = FALSE, names = "unique", null.ok = TRUE)})),
         n_threads               = p_int(lower = 1L, init = 1L, tags = c("train", "catboost", "extra_trees", "fastai", "ft_transformer", "lightgbm", "mlp", "ranger", "resnet", "tabpfn", "xgboost")),
         resampling              = p_uty(init = rsmp("holdout"), tags = c("train", "super"), custom_check = crate({function(x) check_r6(x, "Resampling")})),
         small_data_resampling   = p_uty(init = rsmp("cv", folds = 10), tags = c("train", "super"), custom_check = crate({function(x) check_r6(x, "Resampling")})),